+++ cmd/rcc/main.go
@@ -86,7 +86,7 @@ func markTempForRecycling() {
 	target := common.RobocorpTempName()
 	if pathlib.Exists(target) {
 		filename := filepath.Join(target, "recycle.now")
-		os.WriteFile(filename, []byte("True"), 0o644)
+		pathlib.WriteFile(filename, []byte("True"), 0o644)
 		common.Debug("Marked %q for recycling.", target)
 		markedAlready = true
 	}